requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

python-dotenv==1.0.0
firebase-admin==6.3.0
cloudinary==1.32.0

pydantic>=2.4,<3.0
pydantic-settings>=2.0,<3.0
PyJWT>=2.8,<3.0

Pillow>=10.0.0
opencv-python-headless

numpy==1.26.4
ultralytics==8.1.0

sqlalchemy>=2.0
asyncpg>=0.29.0

anyio>=4.4.0
openai>=1.0.0